Add name suggestion when adding winery or wine
Updated to add name suggestion when adding a new winery or wine. If
user has entered text in search box on SelectWineryActivity or
SelectWineActivity, it will be passed to AddWineryActivity or
AddWineActivity and be displayed in name input. If no search text was
entered then the name input will be blank.