Fix SQL for creating wines table
Fixed issue with SQL to create wines table in DatabaseHelper. It
previously specified that the name column should be unique, which
caused an issue when adding a wine that already exists with a
different vintage or winery. The API will manage any conflicts and
only after a wine is added to the API will it be added to the local
database.